+++ src/modules/src/Eryph.Modules.AspNetCore/ApiProvider/Endpoints/GetEntityEndpoint.cs
@@ -4,27 +4,24 @@ using Eryph.Modules.AspNetCore.ApiProvider.Model;
 using Eryph.StateDb.Model;
 using JetBrains.Annotations;
 
-namespace Eryph.Modules.AspNetCore.ApiProvider.Endpoints
+namespace Eryph.Modules.AspNetCore.ApiProvider.Endpoints;
+
+public abstract class GetEntityEndpoint<TRequest,TResult, TEntity>
+    : SingleResultEndpoint<TRequest, TResult, TEntity> 
+    where TEntity : class 
+    where TRequest : SingleEntityRequest
 {
-    public abstract class GetEntityEndpoint<TRequest,TResult, TEntity> : SingleResultEndpoint<TRequest, TResult, TEntity> 
-        where TEntity : class 
-        where TRequest : SingleEntityRequest
+
+    private readonly ISingleEntitySpecBuilder<TRequest,TEntity> _specBuilder;
+
+    protected GetEntityEndpoint([NotNull] IGetRequestHandler<TEntity, TResult> requestHandler, 
+        ISingleEntitySpecBuilder<TRequest, TEntity> specBuilder) : base(requestHandler)
     {
-
-        private readonly ISingleEntitySpecBuilder<TRequest,TEntity> _specBuilder;
-
-        protected GetEntityEndpoint([NotNull] IGetRequestHandler<TEntity, TResult> requestHandler, 
-            ISingleEntitySpecBuilder<TRequest, TEntity> specBuilder) : base(requestHandler)
-        {
-            _specBuilder = specBuilder;
-        }
-
-        protected override ISingleResultSpecification<TEntity> CreateSpecification(TRequest request)
-        {
-            return _specBuilder.GetSingleEntitySpec(request, AccessRight.Read);
-        }
-
-        
+        _specBuilder = specBuilder;
     }
 
+    protected override ISingleResultSpecification<TEntity> CreateSpecification(TRequest request)
+    {
+        return _specBuilder.GetSingleEntitySpec(request, AccessRight.Read);
+    }
 }
