+++ src/modules/src/Eryph.Modules.AspNetCore/ApiProvider/Endpoints/SingleResultEndpoint.cs
@@ -4,31 +4,31 @@ using Ardalis.ApiEndpoints;
 using Ardalis.Specification;
 using Eryph.Modules.AspNetCore.ApiProvider.Handlers;
 using Eryph.Modules.AspNetCore.ApiProvider.Model;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Eryph.Modules.AspNetCore.ApiProvider.Endpoints
+namespace Eryph.Modules.AspNetCore.ApiProvider.Endpoints;
+
+[Route("v{version:apiVersion}")]
+public abstract class SingleResultEndpoint<TRequest, TResult, TModel> : EndpointBaseAsync
+    .WithRequest<TRequest>
+    .WithActionResult<TResult> 
+    where TModel : class
+    where TRequest: RequestBase
 {
-    [Route("v{version:apiVersion}")]
-    public abstract class SingleResultEndpoint<TRequest, TResult, TModel> : EndpointBaseAsync
-        .WithRequest<TRequest>
-        .WithActionResult<TResult> 
-        where TModel : class
-        where TRequest: RequestBase
+    private readonly IGetRequestHandler<TModel, TResult> _requestHandler;
+    
+    protected SingleResultEndpoint(IGetRequestHandler<TModel, TResult> requestHandler)
     {
-        private readonly IGetRequestHandler<TModel, TResult> _requestHandler;
-        public SingleResultEndpoint(IGetRequestHandler<TModel, TResult> requestHandler)
-        {
-            _requestHandler = requestHandler;
-        }
-
-        protected abstract ISingleResultSpecification<TModel> CreateSpecification(TRequest request);
-
-        public override Task<ActionResult<TResult>> HandleAsync(TRequest request, CancellationToken cancellationToken = default)
-        {
-            return _requestHandler.HandleGetRequest(
-                () => CreateSpecification(request), cancellationToken );
-        }
+        _requestHandler = requestHandler;
     }
 
+    protected abstract ISingleResultSpecification<TModel> CreateSpecification(TRequest request);
+
+    public override Task<ActionResult<TResult>> HandleAsync(
+        TRequest request,
+        CancellationToken cancellationToken = default)
+    {
+        return _requestHandler.HandleGetRequest(
+            () => CreateSpecification(request), cancellationToken );
+    }
 }
