+++ src/modules/src/Eryph.Modules.AspNetCore/ApiProvider/Handlers/IOperationRequestHandler.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Eryph.Modules.AspNetCore.ApiProvider.Model;
 using Eryph.Modules.AspNetCore.ApiProvider.Model.V1;
 using Microsoft.AspNetCore.Mvc;
 
@@ -12,7 +11,7 @@ namespace Eryph.Modules.AspNetCore.ApiProvider.Handlers;
 
 public interface IOperationRequestHandler<TEntity> where TEntity : class
 {
-    Task<ActionResult<ListEntitiesResponse<Operation>>> HandleOperationRequest(
+    Task<ActionResult<Operation>> HandleOperationRequest(
         Func<object> createOperationFunc,
         CancellationToken cancellationToken);
 }
